feat(blogpost): Require login for create, edit or delete blogposts

+++ db/models/blogpost.py
@@ -4,6 +4,7 @@ from sqlalchemy.sql import func
 from fastapi import status
 
 from db import Base
+from db.models.user import User
 from schemas import BlogpostCreate, BlogpostUpdate
 
 
@@ -40,8 +41,8 @@ def retrieve_post_list(db: Session):
     )
 
 
-def create_new_blogpost(blogpost: BlogpostCreate, db: Session):
-    new_blogpost = Blogpost(**blogpost.__dict__, author_id=5)
+def create_new_blogpost(blogpost: BlogpostCreate, author_id: int, db: Session):
+    new_blogpost = Blogpost(**blogpost.__dict__, author_id=author_id)
     db.add(new_blogpost)
     db.commit()
     return retrieve_blogpost_by_slug(new_blogpost.slug, db)
@@ -65,12 +66,20 @@ def retrieve_blogpost_by_slug(slug: str, db: Session):
     )
 
 
-def update_blogpost_with_id(post_id: int, new_data: BlogpostUpdate, db: Session):
+def update_blogpost_with_id(
+    post_id: int, new_data: BlogpostUpdate, current_user: User, db: Session
+):
     old_post = retrieve_blogpost_by_id(post_id, db)
     if not old_post:
         return {
             "status_code": status.HTTP_404_NOT_FOUND,
             "detail": f"Post with ID {post_id} does not exists.",
+        }
+
+    if not current_user.is_superuser and current_user.id != old_post.author_id:
+        return {
+            "status_code": status.HTTP_403_FORBIDDEN,
+            "detail": "You are not allowed to edit this post.",
         }
 
     for key, value in new_data.__dict__.items():
@@ -80,12 +89,21 @@ def update_blogpost_with_id(post_id: int, new_data: BlogpostUpdate, db: Session)
     return retrieve_blogpost_by_id(post_id, db)
 
 
-def delete_blogpost_by_id(post_id: int, db: Session):
+def delete_blogpost_by_id(post_id: int, current_user: User, db: Session):
     post_to_be_deleted = db.query(Blogpost).filter(Blogpost.id == post_id)
     if not post_to_be_deleted.one_or_none():
         return {
             "status_code": status.HTTP_404_NOT_FOUND,
             "detail": f"Post with ID {post_id} does not exists.",
+        }
+
+    if (
+        not current_user.is_superuser
+        and current_user.id != post_to_be_deleted.one_or_none().author_id
+    ):
+        return {
+            "status_code": status.HTTP_403_FORBIDDEN,
+            "detail": "You are not allowed to delete this post.",
         }
 
     post_to_be_deleted.delete()
